Mutualise les règles communes du parser
